Create README.md
Functional Dichotomy of Orbitofrontal Cortex and Anterior Cingulate Cortex Subregions in Decision-Making and Brain-Body Regulation

This repository contains MATLAB and Python scripts used for the analysis, preprocessing, and modeling of physiological and neural data, supporting the findings of the manuscript submitted to Nature Communications. The study investigates the distinct functional roles of subregions in the orbitofrontal cortex (OFC) and anterior cingulate cortex (ACC) in decision-making processes and brain-body regulation.